remove PathOps
