+++ branch/src/main/scala/dev/alteration/branch/spider/server/FileContextHandler.scala
@@ -1,15 +1,39 @@
 package dev.alteration.branch.spider.server
 
 import com.sun.net.httpserver.{Authenticator, Filter}
-import dev.alteration.branch.macaroni.fs.PathOps
 import dev.alteration.branch.spider.*
-import dev.alteration.branch.macaroni.fs.PathOps.*
-import dev.alteration.branch.spider.HttpMethod
 
 import java.io.File
 import java.nio.file.{Files, Path}
 
 object FileContextHandler {
+
+  private def pathFromStrList(segments: List[String]): Path = {
+    segments match {
+      case Nil          => Path.of("")
+      case head :: Nil  => Path.of(head)
+      case head :: tail => tail.foldLeft(Path.of(head))(_ / _)
+    }
+  }
+
+  extension (path: Path) {
+
+    def /(segment: String | Path): Path = {
+      if (segment == null)
+        throw new IllegalArgumentException("Path segment cannot be null")
+      segment match {
+        case path: Path  => path.resolve(path)
+        case str: String => path.resolve(str)
+      }
+    }
+
+    def relativeTo(rootPath: Path): Path = {
+      if (rootPath == null)
+        throw new IllegalArgumentException("Root path cannot be null")
+      rootPath.relativize(path)
+    }
+
+  }
 
   /** A list of default files to look for when a directory is requested. E.g.
     * /some/path -> /some/path/index.html
@@ -37,6 +61,7 @@ case class FileContextHandler(
     override val filters: Seq[Filter] = Seq.empty,
     override val authenticator: Option[Authenticator] = Option.empty
 ) extends ContextHandler(contextPath) {
+  import FileContextHandler.*
 
   private def fileExists(path: Path): Boolean = {
     val filePath = (rootFilePath / path).toString
@@ -77,10 +102,13 @@ case class FileContextHandler(
 
   override val contextRouter
       : PartialFunction[(HttpMethod, List[String]), RequestHandler[?, ?]] = {
-    case HttpMethod.GET -> anyPath if fileExists(PathOps.fromList(anyPath)) =>
+    case HttpMethod.GET -> anyPath
+        if fileExists(FileContextHandler.pathFromStrList(anyPath)) =>
       fileHandler
     case HttpMethod.GET -> anyPath
-        if defaultExists(PathOps.fromList(anyPath)) =>
-      DefaultFileHandler(defaultFile(PathOps.fromList(anyPath)))
+        if defaultExists(FileContextHandler.pathFromStrList(anyPath)) =>
+      DefaultFileHandler(
+        defaultFile(FileContextHandler.pathFromStrList(anyPath))
+      )
   }
 }
