+++ branch/src/main/scala/dev/alteration/branch/spider/server/FileHandler.scala
@@ -1,6 +1,5 @@
 package dev.alteration.branch.spider.server
 
-import dev.alteration.branch.macaroni.fs.PathOps.*
 import FileHandler.given
 import RequestHandler.given
 
@@ -31,8 +30,9 @@ private[spider] case class FileHandler(rootFilePath: Path)
     extends RequestHandler[Unit, File] {
 
   override def handle(request: Request[Unit]): Response[File] = {
-    val filePath =
-      (rootFilePath / request.uri.getPath.stripPrefix("/")).toString
+    val filePath = {
+      rootFilePath.resolve(request.uri.getPath.stripPrefix("/")).toString
+    }
     Response(
       200,
       body = new File(filePath)
