+++ branch/src/test/scala/dev/alteration/branch/spider/client/ClientSpec.scala
@@ -5,8 +5,8 @@ import ClientRequest.uri
 import dev.alteration.branch.friday.JsonEncoder
 import dev.alteration.branch.friday.http.JsonBodyHandler
 import dev.alteration.branch.spider.server.RequestHandler.given
-import dev.alteration.branch.macaroni.fs.PathOps.*
 import dev.alteration.branch.spider.HttpMethod
+import dev.alteration.branch.spider.server.ContextHandler.ci
 import dev.alteration.branch.spider.server.{Request, RequestHandler, Response}
 import munit.FunSuite
