+++ branch/src/test/scala/dev/alteration/branch/spider/server/FileContextHandlerSpec.scala
@@ -1,12 +1,13 @@
 package dev.alteration.branch.spider.server
 
 import dev.alteration.branch.testkit.fixtures.FileFixtureSuite
-import dev.alteration.branch.macaroni.fs.PathOps.*
 import dev.alteration.branch.spider.HttpMethod
 
 import java.nio.file.Files
 
 class FileContextHandlerSpec extends FileFixtureSuite {
+
+  import FileContextHandler.*
 
   tmpDir.test("FileContextHandler") { tmpDir =>
